Steve AI emotion regulation

+++ src/main/java/net/notccg/yahresurrected/entity/custom/logic/behaviors/EmotionControlBehaviour.java
@@ -4,7 +4,6 @@ import com.mojang.datafixers.util.Pair;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.PathfinderMob;
-import net.minecraft.world.entity.ai.Brain;
 import net.minecraft.world.entity.ai.memory.MemoryModuleType;
 import net.minecraft.world.entity.ai.memory.MemoryStatus;
 import net.notccg.yahresurrected.entity.custom.logic.steve_ai.SteveLogic;
@@ -15,13 +14,13 @@ import java.util.List;
 
 public class EmotionControlBehaviour<E extends PathfinderMob> extends ExtendedBehaviour<E> {
     private final long emotionRundownTicks;
-    private final float fearDecreaseMultiplier;
-    private final float curiosityDecreaseMultiplier;
-    private final float paranoiaDecreaseMultiplier;
+    private final double fearDecreaseMultiplier;
+    private final double curiosityDecreaseMultiplier;
+    private final double paranoiaDecreaseMultiplier;
 
     private long nextRundownInterval = 0;
 
-    public EmotionControlBehaviour(long emotionRundownTicks, float fearDecreaseMultiplier, float curiosityDecreaseMultiplier, float paranoiaDecreaseMultiplier) {
+    public EmotionControlBehaviour(long emotionRundownTicks, double fearDecreaseMultiplier, double curiosityDecreaseMultiplier, double paranoiaDecreaseMultiplier) {
         this.emotionRundownTicks = emotionRundownTicks;
         this.fearDecreaseMultiplier = fearDecreaseMultiplier;
         this.curiosityDecreaseMultiplier = curiosityDecreaseMultiplier;
